Extract event related functions into own file
All event related code can now be found in event.go
