+++ pkg/master/master.go
@@ -11,7 +11,6 @@ import (
 	"strings"
 	"sync"
 	"sync/atomic"
-	"time"
 
 	"golang.org/x/net/context"
 
@@ -31,14 +30,6 @@ func init() {
 		PhaseIDReport:   newReportPhase,
 	}
 }
-
-type EventClass string
-
-const (
-	EventAll    EventClass = "all"
-	EventPhase  EventClass = "phase"
-	EventModule EventClass = "module"
-)
 
 type server struct {
 	analysisClosed     chan bool
@@ -159,61 +150,6 @@ func (s *server) Quit(ctx context.Context, in *service.QuitMessage) (*service.Qu
 	quitServer <- nil
 
 	return &service.QuitResponse{Success: true}, nil
-}
-
-func (s *server) SubscribeEvents(in *service.EventMessage, stream service.ControlService_SubscribeEventsServer) error {
-	var eventKey EventClass
-	switch in.Class {
-	case "all":
-		eventKey = EventAll
-	case "phase":
-		eventKey = EventPhase
-	case "module":
-		eventKey = EventModule
-	default:
-		return fmt.Errorf("No such event class %s", in.Class)
-	}
-	eventC := make(chan *service.Event)
-	err := s.registerEventChannel(eventKey, eventC)
-	if err != nil {
-		return err
-	}
-
-	for event := range eventC {
-		stream.Send(event)
-	}
-	return nil
-}
-
-func (s *server) registerEventChannel(eventClass EventClass, eventChannel chan *service.Event) error {
-	if _, ok := s.eventChannels[eventClass]; !ok {
-		return fmt.Errorf("No such event class %v", eventClass)
-	}
-	s.eventChannels[eventClass] = append(s.eventChannels[eventClass], eventChannel)
-	return nil
-}
-
-func (s *server) publishEvent(event *service.Event) error {
-	if _, ok := s.eventChannels[EventClass(event.Class)]; !ok {
-		return fmt.Errorf("No such event class %v", event.Class)
-	}
-	for _, sub := range []EventClass{EventClass(event.Class), EventAll} {
-		for _, channel := range s.eventChannels[sub] {
-			go func(c chan *service.Event) {
-				ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
-				defer cancel()
-				select {
-				case c <- event:
-					return
-				case <-ctx.Done():
-					log.Printf("Could not send event %v to channel", event)
-				}
-			}(channel)
-		}
-	}
-
-	// TODO publish the events to the corresponding channels
-	return nil
 }
 
 // InitAndRun sets up and runs the grpc services and the dgraph database connection
